Make the Feather connection to WiFi more stable.
When the WiFi is disconnected, for whatever reason, the next request will
fail with a cryptic error. To avoid leaving the wrong notification on, if
that particular error is received, try re-connecting to WiFi.

(It would be easier to reboot, but that doesn't seem safe when connected to
a computer.)

Also make it a bit easier to see when the WiFi is connecting, by changing
the colour of the light slightly.