Improve error handling for the Feather.
After any three exceptions are caught, do a soft reboot of the board by
reloading the supervisor. This allows clearing error statuses without
sticking to the wrong LED colour.

If there is no led attribute in the JSON response (for instance because the
server returned 404), use a white(ish) colour to signifiy the error
condition, and treat it as an exception.